+++ NEWS.md
@@ -1,14 +1,18 @@
-# lmhelprs 0.4.5.2
+# lmhelprs 0.4.5.3
 
 ## New Features
 
-- Added an `update()` method to the output
+- Added an `update()` method for the output
   of `many_lm()`.
   (0.4.5.1)
 
-- Added a `coef()` method to the output
+- Added a `coef()` method for the output
   of `many_lm()`.
   (0.4.5.2)
+
+- Added a `vcov()` method for the output
+  of `many_lm()`.
+  (0.4.5.3)
 
 # lmhelprs 0.4.5